feature: Refactor package structure and update imports for AI components
