+++ ai-agent-bedrock/src/main/scala/wvlet/ai/agent/chat/bedrock/BedrockClient.scala
@@ -1,0 +1,37 @@
+package wvlet.ai.agent.chat.bedrock
+
+import software.amazon.awssdk.services.bedrockruntime.BedrockRuntimeAsyncClient
+import software.amazon.awssdk.services.bedrockruntime.model.{
+  ConverseStreamRequest,
+  ConverseStreamResponseHandler
+}
+import wvlet.ai.agent.chat.{ChatObserver, ChatRequest}
+
+import java.util.concurrent.CompletableFuture
+
+trait BedrockClient extends AutoCloseable:
+  def converseStream(
+      request: ConverseStreamRequest,
+      handler: ConverseStreamResponseHandler
+  ): CompletableFuture[Void]
+
+object BedrockClient:
+  def apply(config: BedrockConfig): BedrockClient = BedrockAsyncClient(config)
+  def apply(): BedrockClient                      = apply(BedrockConfig())
+
+class BedrockAsyncClient(config: BedrockConfig) extends BedrockClient:
+  import wvlet.ai.agent.core.ops.*
+
+  private val client = BedrockRuntimeAsyncClient
+    .builder()
+    .region(config.region)
+    .credentialsProvider(config.credentialProvider)
+    .pipe(config.asyncClientConfig(_))
+    .build()
+
+  override def converseStream(
+      request: ConverseStreamRequest,
+      handler: ConverseStreamResponseHandler
+  ): CompletableFuture[Void] = client.converseStream(request, handler)
+
+  override def close(): Unit = client.close()
