+++ ai-agent-bedrock/src/main/scala/wvlet/ai/agent/chat/bedrock/DocumentUtil.scala
