+++ ai-agent-bedrock/src/test/scala/wvlet/ai/agent/chat/bedrock/BedrockIntegrationTest.scala
