+++ ai-agent/src/main/scala/wvlet/ai/agent/chat/ChatContent.scala
