+++ ai-agent/src/main/scala/wvlet/ai/agent/chat/ChatMessage.scala
