+++ ai-agent/src/main/scala/wvlet/ai/agent/chat/ChatResponse.scala
