+++ ai-agent/src/main/scala/wvlet/ai/agent/core/AIException.scala
@@ -1,0 +1,4 @@
+package wvlet.ai.agent.core
+
+case class AIException(statusCode: StatusCode, message: String, cause: Throwable = null)
+    extends Exception(message, cause)
