--- ai-core/src/main/scala/wvlet/ai/agent/chat/ToolSpec.scala
@@ -1,17 +1,0 @@
-package wvlet.ai.agent.chat
-
-import wvlet.ai.core.DataType
-
-case class ToolSpec(
-    name: String,
-    description: String,
-    parameters: List[ToolParameter],
-    returnType: DataType
-)
-
-case class ToolParameter(
-    name: String,
-    description: String,
-    dataType: DataType,
-    defaultValue: Option[Any] = None
-)
