--- ai-core/src/main/scala/wvlet/ai/core/AIException.scala
@@ -1,4 +1,0 @@
-package wvlet.ai.core
-
-case class AIException(statusCode: StatusCode, message: String, cause: Throwable = null)
-    extends Exception(message, cause)
